# Update stable_hash usage to 1.0