--- a/NEWS.md
+++ b/NEWS.md
@@ -6,6 +6,7 @@
 - Improved Barplot Label Alignment [#3160](https://github.com/MakieOrg/Makie.jl/issues/3160).
 - Fixed regression in determining axis limits [#3179](https://github.com/MakieOrg/Makie.jl/pull/3179)
 - Added a theme `theme_latexfonts` that uses the latex font family as default fonts [#3147](https://github.com/MakieOrg/Makie.jl/pull/3147).
+- Upgrades `StableHashTraits` from 0.3 to 1.0
 
 ## v0.19.8