--- a/src/bezier.jl
+++ b/src/bezier.jl
@@ -42,13 +42,6 @@ struct BezierPath
     commands::Vector{PathCommand}
 end
 
-StableHashTraits.transform(path::BezierPath) = path.commands
-StableHashTraits.transform(c::EllipticalArc) = [c.c[1], c.c[2], c.r1, c.r2, c.angle, c.a1, c.a2]
-StableHashTraits.transform(c::CurveTo) = [c.c1[1], c.c1[2], c.c2[1], c.c2[2], c.p[1], c.p[2]]
-StableHashTraits.transform(c::LineTo) = [c.p[1], c.p[2]]
-StableHashTraits.transform(c::MoveTo) = [c.p[1], c.p[2]]
-StableHashTraits.transform(c::ClosePath) = 0
-
 # so that the same bezierpath with a different instance of a vector hashes the same
 # and we don't create the same texture atlas entry twice
 Base.:(==)(b1::BezierPath, b2::BezierPath) = b1.commands == b2.commands