--- a/src/utilities/texture_atlas.jl
+++ b/src/utilities/texture_atlas.jl
@@ -292,14 +292,16 @@ function glyph_uv_width!(atlas::TextureAtlas, b::BezierPath)
     return atlas.uv_rectangles[glyph_index!(atlas, b)]
 end
 
+crc(x, seed=UInt32(0)) = crc32c(collect(x), seed)
 function insert_glyph!(atlas::TextureAtlas, glyph, font::NativeFont)
     glyphindex = FreeTypeAbstraction.glyph_index(font, glyph)
-    hash = StableHashTraits.stable_hash((glyphindex, FreeTypeAbstraction.fontname(font)))
+    hash = StableHashTraits.stable_hash((glyphindex, FreeTypeAbstraction.fontname(font));
+                                        alg=crc)
     return insert_glyph!(atlas, hash, (glyphindex, font))
 end
 
 function insert_glyph!(atlas::TextureAtlas, path::BezierPath)
-    return insert_glyph!(atlas, StableHashTraits.stable_hash(path), path)
+    return insert_glyph!(atlas, StableHashTraits.stable_hash(path; alg=crc), path)
 end