--- a/src/utilities/texture_atlas.jl
+++ b/src/utilities/texture_atlas.jl
@@ -1,3 +1,4 @@
+const SERIALIZATION_FORMAT_VERSION = "v4"
 
 struct TextureAtlas
     rectangle_packer::RectanglePacker{Int32}
@@ -69,8 +70,6 @@ function Base.show(io::IO, atlas::TextureAtlas)
     println(io, "  downsample: ", atlas.downsample)
     println(io, "  font_render_callback: ", length(atlas.font_render_callback))
 end
-
-const SERIALIZATION_FORMAT_VERSION = "v3"
 
 # basically a singleton for the textureatlas
 function get_cache_path(resolution::Int, pix_per_glyph::Int)