# Description

This upgrades `StableHashTraits` to 1.0 and revises its usage accordingly.

## Type of change

Delete options that do not apply:

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist

- [ ] Added an entry in NEWS.md (for new features and breaking changes)
- [ ] ~~Added or changed relevant sections in the documentation~~
- [ ] ~~Added unit tests for new algorithms, conversion methods, etc.~~
- [ ] ~~Added reference image tests for new plotting functions, recipes, visual options, etc.~~