@SimonDanisch 2023-09-06
Any idea how to fix the performance regression?
This PR:
```julia
path = Makie.bezier_star(8, 0.33, 0.6, pi/2
julia> @btime Makie.StableHashTraits.stable_hash(path; alg=Makie.crc)
  131.000 μs (4977 allocations: 254.51 KiB)
julia> @btime Makie.StableHashTraits.stable_hash(('c', Makie.FreeTypeAbstraction.fontname(Makie.defaultfont())); alg=Makie.crc)
  2.511 μs (100 allocations: 5.20 KiB)
```
Master:
```julia
julia> @btime Makie.StableHashTraits.stable_hash(path)
  5.733 μs (366 allocations: 16.54 KiB)
julia> @btime Makie.StableHashTraits.stable_hash(('c', Makie.FreeTypeAbstraction.fontname(Makie.defaultfont())))
  695.652 ns (28 allocations: 1.30 KiB)
```