@haberdashPI 2023-09-06
The new version is slower, so this may just be the timings you get with the updates. 

I have some performance improvements in the works: (see https://github.com/beacon-biosignals/StableHashTraits.jl/pull/29 and https://github.com/beacon-biosignals/StableHashTraits.jl/pull/30), which should land as a part of v1.1.