@SimonDanisch 2023-09-07
Merged in https://github.com/MakieOrg/Makie.jl/pull/3216